renamed models and visualisation engine to be standardised. Fixed import statements for WLS, HMM, DTMC models.
